requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9       # needed for UploadFile

# ML
tensorflow==2.15.0            # or tensorflow-cpu if no GPU
numpy==1.26.4
Pillow==10.3.0

# Data
pandas==2.2.2

matplotlib==3.9.0

# Dev
httpx==0.27.0                 # for testing with FastAPI TestClient
